V0.3
Updated the `Update()` method to check if the player is within a certain distance of the edges of the terrain before adjusting the terrain detail settings. This will prevent the terrain detail settings from being adjusted at the edges of the terrain, which should help to hide any seams between terrains.